business_owner_ai/requirements.txt
# ================================================================
# ndara.ai Business Owner AI - Requirements
# ================================================================

# Build / base (CVE-2022-40897 ReDoS, CVE-2024-6345, CVE-2025-47273)
setuptools>=78.1.1

# Core AI/ML
openai>=1.12.0                    # OpenAI GPT models (SDK 2.x compatible)
tiktoken==0.5.2                   # Token counting

# Machine Learning (for advanced features)
scikit-learn>=1.3.0               # Clustering, segmentation
scipy>=1.11.0                     # Scientific computing
numpy==1.26.4                     # Numerical operations

# Data Processing
pydantic==2.5.0                   # Data validation
python-dotenv==1.0.0              # Environment management

# API Framework (Minimal) - FastAPI>=0.116.0 pulls starlette>=0.40.0 (CVE-2024-47874)
fastapi>=0.116.0                  # Lightweight API
uvicorn[standard]==0.24.0         # ASGI server

# Utilities
aiofiles==23.2.0                  # Async file operations
python-multipart>=0.0.22          # Form data (CVE-2024-24762, CVE-2024-53981, CVE-2026-24486)
httpx==0.25.2                     # HTTP client for appointment service

# Optional: Advanced Analytics
# pandas>=2.0.0                   # Data analysis
# plotly>=5.0.0                   # Visualization

# Development Tools (Optional)
# pytest==7.4.3                   # Testing
# black==23.7.0                   # Code formatting
# flake8==6.0.0                   # Linting

